requirements_vec4ir.txt
# Requirements specified in setup.py
boto
bz2file
cycler
decorator
gensim
isodate
matplotlib
networkx
nltk
pandas
PyYAML
rdflib
requests
scikit-learn
scipy
six
sklearn
smart-open
joblib
tensorflow
numpy

# Development requirements

pytest
pylint
jedi
PyJWT

